add loging setting

+++ 01_gen_data/rollout.sh
@@ -98,7 +98,9 @@ function gen_data() {
     for h in $(psql ${PSQL_OPTIONS} -v ON_ERROR_STOP=1 -q -A -t -c "${SQL_QUERY}"); do
       EXT_HOST=$(echo ${h} | awk -F '|' '{print $2}')
       SEG_DATA_PATH=$(echo ${h} | awk -F '|' '{print $3}' | sed 's#//#/#g')
-      log_time "ssh -n ${EXT_HOST} \"rm -rf ${SEG_DATA_PATH}/hbenchmark; mkdir -p ${SEG_DATA_PATH}/hbenchmark/logs\" &"
+      if [ "${LOG_DEBUG}" == "true" ]; then
+        log_time "ssh -n ${EXT_HOST} \"rm -rf ${SEG_DATA_PATH}/hbenchmark; mkdir -p ${SEG_DATA_PATH}/hbenchmark/logs\" &"
+      fi
       ssh -n ${EXT_HOST} "rm -rf ${SEG_DATA_PATH}/hbenchmark; mkdir -p ${SEG_DATA_PATH}/hbenchmark/logs" &
     done
     wait
@@ -109,7 +111,9 @@ function gen_data() {
       SEG_DATA_PATH=$(echo ${i} | awk -F '|' '{print $3}' | sed 's#//#/#g')
       for ((j=1; j<=GEN_DATA_PARALLEL; j++)); do
         GEN_DATA_PATH="${SEG_DATA_PATH}/hbenchmark/${CHILD}"
-        log_time "ssh -n ${EXT_HOST} \"bash -c 'cd ~/; ./generate_data.sh ${GEN_DATA_SCALE} ${CHILD} ${PARALLEL} ${GEN_DATA_PATH} > ${SEG_DATA_PATH}/hbenchmark/logs/tpch.generate_data.${CHILD}.log 2>&1 &'\""
+        if [ "${LOG_DEBUG}" == "true" ]; then
+          log_time "ssh -n ${EXT_HOST} \"bash -c 'cd ~/; ./generate_data.sh ${GEN_DATA_SCALE} ${CHILD} ${PARALLEL} ${GEN_DATA_PATH} > ${SEG_DATA_PATH}/hbenchmark/logs/tpch.generate_data.${CHILD}.log 2>&1 &'\""
+        fi
         ssh -n ${EXT_HOST} "bash -c 'cd ~/; ./generate_data.sh ${GEN_DATA_SCALE} ${CHILD} ${PARALLEL} ${GEN_DATA_PATH} > ${SEG_DATA_PATH}/hbenchmark/logs/tpch.generate_data.${CHILD}.log 2>&1 &'" &
         CHILD=$((CHILD + 1))
       done
@@ -139,7 +143,9 @@ function gen_data() {
     for EXT_HOST in $(cat ${TPC_H_DIR}/segment_hosts.txt); do
       # For each path, start a gpfdist instance
       for GEN_DATA_PATH in "${GEN_PATHS[@]}"; do
-        log_time "ssh -n ${EXT_HOST} \"rm -rf ${GEN_DATA_PATH}/hbenchmark; mkdir -p ${GEN_DATA_PATH}/hbenchmark/logs\" &"
+        if [ "${LOG_DEBUG}" == "true" ]; then
+          log_time "ssh -n ${EXT_HOST} \"rm -rf ${GEN_DATA_PATH}/hbenchmark; mkdir -p ${GEN_DATA_PATH}/hbenchmark/logs\" &"
+        fi
         ssh -n ${EXT_HOST} "rm -rf ${GEN_DATA_PATH}/hbenchmark; mkdir -p ${GEN_DATA_PATH}/hbenchmark/logs" &
       done
     done
@@ -151,7 +157,9 @@ function gen_data() {
       for GEN_DATA_PATH in "${GEN_PATHS[@]}"; do
         for ((j=1; j<=GEN_DATA_PARALLEL; j++)); do
           GEN_DATA_SUBPATH="${GEN_DATA_PATH}/hbenchmark/${CHILD}"
-          log_time "ssh -n ${EXT_HOST} \"bash -c 'cd ~/; ./generate_data.sh ${GEN_DATA_SCALE} ${CHILD} ${PARALLEL} ${GEN_DATA_SUBPATH} > ${GEN_DATA_PATH}/hbenchmark/logs/tpch.generate_data.${CHILD}.log 2>&1 &'\""
+          if [ "${LOG_DEBUG}" == "true" ]; then
+            log_time "ssh -n ${EXT_HOST} \"bash -c 'cd ~/; ./generate_data.sh ${GEN_DATA_SCALE} ${CHILD} ${PARALLEL} ${GEN_DATA_SUBPATH} > ${GEN_DATA_PATH}/hbenchmark/logs/tpch.generate_data.${CHILD}.log 2>&1 &'\""
+          fi
           ssh -n ${EXT_HOST} "bash -c 'cd ~/; ./generate_data.sh ${GEN_DATA_SCALE} ${CHILD} ${PARALLEL} ${GEN_DATA_SUBPATH} > ${GEN_DATA_PATH}/hbenchmark/logs/tpch.generate_data.${CHILD}.log 2>&1 &'" &
           CHILD=$((CHILD + 1))
         done
@@ -191,12 +199,18 @@ if [ "${GEN_NEW_DATA}" == "true" ]; then
     # Prepare each data generation path
     for GEN_DATA_PATH in "${GEN_PATHS[@]}"; do
       if [[ ! -d "${GEN_DATA_PATH}" && ! -L "${GEN_DATA_PATH}" ]]; then
-        log_time "mkdir ${GEN_DATA_PATH}/hbenchmark"
+        if [ "${LOG_DEBUG}" == "true" ]; then
+          log_time "mkdir ${GEN_DATA_PATH}/hbenchmark"
+        fi
         mkdir -p ${GEN_DATA_PATH}/hbenchmark
       fi
-      log_time "rm -rf ${GEN_DATA_PATH}/hbenchmark/*"
+      if [ "${LOG_DEBUG}" == "true" ]; then
+        log_time "rm -rf ${GEN_DATA_PATH}/hbenchmark/*"
+      fi
       rm -rf ${GEN_DATA_PATH}/hbenchmark/*
-      log_time "mkdir -p ${GEN_DATA_PATH}/hbenchmark/logs"
+      if [ "${LOG_DEBUG}" == "true" ]; then
+        log_time "mkdir -p ${GEN_DATA_PATH}/hbenchmark/logs"
+      fi
       mkdir -p ${GEN_DATA_PATH}/hbenchmark/logs
     done
 
@@ -204,7 +218,9 @@ if [ "${GEN_NEW_DATA}" == "true" ]; then
     for GEN_DATA_PATH in "${GEN_PATHS[@]}"; do
       for ((j=1; j<=GEN_DATA_PARALLEL; j++)); do
         GEN_DATA_SUBPATH="${GEN_DATA_PATH}/hbenchmark/${CHILD}"
-        log_time "sh ${TPC_H_DIR}/01_gen_data/generate_data.sh ${GEN_DATA_SCALE} ${CHILD} ${PARALLEL} ${GEN_DATA_SUBPATH} > ${GEN_DATA_PATH}/hbenchmark/logs/tpch.generate_data.${CHILD}.log 2>&1 &"
+        if [ "${LOG_DEBUG}" == "true" ]; then
+          log_time "sh ${TPC_H_DIR}/01_gen_data/generate_data.sh ${GEN_DATA_SCALE} ${CHILD} ${PARALLEL} ${GEN_DATA_SUBPATH} > ${GEN_DATA_PATH}/hbenchmark/logs/tpch.generate_data.${CHILD}.log 2>&1 &"
+        fi
         sh ${TPC_H_DIR}/01_gen_data/generate_data.sh ${GEN_DATA_SCALE} ${CHILD} ${PARALLEL} ${GEN_DATA_SUBPATH} > ${GEN_DATA_PATH}/hbenchmark/logs/tpch.generate_data.${CHILD}.log 2>&1 &
         CHILD=$((CHILD + 1))
       done
